+++ crd-generator/api-v2/src/main/java/io/fabric8/crdv2/generator/CRDGenerator.java
@@ -47,6 +47,8 @@ import java.util.stream.Collectors;
 public class CRDGenerator {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(CRDGenerator.class);
+  private static final CRDPostProcessor nullProcessor = new CRDPostProcessor() {
+  };
   private final Map<String, AbstractCustomResourceHandler> handlers = new HashMap<>(2);
   private CRDOutput<? extends OutputStream> output;
   private boolean parallel;
@@ -156,7 +158,7 @@ public class CRDGenerator {
     return detailedGenerate().numberOfGeneratedCRDs();
   }
 
-  public CRDGenerationInfo detailedGenerate() {
+  public CRDGenerationInfo detailedGenerate(CRDPostProcessor processor) {
     if (getCustomResourceInfos().isEmpty()) {
       LOGGER.warn("No resources were registered with the 'customResources' method to be generated");
       return CRDGenerationInfo.EMPTY;
@@ -192,22 +194,40 @@ public class CRDGenerator {
 
         if (parallel) {
           handlers.values().stream().map(h -> ForkJoinPool.commonPool().submit(() -> h.handle(info, context.forkContext())))
-              .collect(Collectors.toList()).stream().forEach(ForkJoinTask::join);
+              .collect(Collectors.toList())
+              .forEach(ForkJoinTask::join);
         } else {
-          handlers.values().stream().forEach(h -> h.handle(info, context.forkContext()));
+          handlers.values().forEach(h -> h.handle(info, context.forkContext()));
         }
       }
     }
 
     final CRDGenerationInfo crdGenerationInfo = new CRDGenerationInfo();
     handlers.values().stream().flatMap(AbstractCustomResourceHandler::finish)
-        .forEach(crd -> emitCrd(crd.getKey(), crd.getValue(), crdGenerationInfo));
+        .forEach(crd -> emitCrd(crd.getKey(), crd.getValue(), crdGenerationInfo, processor));
     return crdGenerationInfo;
   }
 
+  public CRDGenerationInfo detailedGenerate() {
+    return detailedGenerate(nullProcessor);
+  }
+
+  /**
+   * @deprecated use {@link #emitCrd(HasMetadata, Set, CRDGenerationInfo, CRDPostProcessor)} instead
+   */
+  @Deprecated(forRemoval = true)
   public void emitCrd(HasMetadata crd, Set<String> dependentClassNames, CRDGenerationInfo crdGenerationInfo) {
+    emitCrd(crd, dependentClassNames, crdGenerationInfo, nullProcessor);
+  }
+
+  protected void emitCrd(HasMetadata crd, Set<String> dependentClassNames, CRDGenerationInfo crdGenerationInfo,
+      CRDPostProcessor processor) {
     final String version = ApiVersionUtil.trimVersion(crd.getApiVersion());
     final String crdName = crd.getMetadata().getName();
+
+    // post-process the CRD if needed
+    crd = processor.process(crd, version);
+
     try {
       final String outputName = getOutputName(crdName, version);
       try (final OutputStreamWriter writer = new OutputStreamWriter(output.outputFor(outputName), StandardCharsets.UTF_8)) {